Mark NPCs as quest givers, add conversations to Quests
Mark them as quest givers, to eventually be able to selectively display
a `(!)` next to their name if they have a quest. We want to only perform
this check if we know they are a quest giver. Add conversations to
quests so that, if they are a quest giver, we can find the next
available quest and use that conversation as a replacement.