Fix NPCs getting an endless loop of combat ticks
If a player respawned in the same room as the NPC, they would act on the
message which triggers a combat tick, which will kill the PC. The NPC
still has their target, the PC respawns into that room trigger another
combat tick, and the old one is still valid. Rinse repeat until the NPC
has 10+ combat ticks going.

Prevent this by only matching the target action when combat is not
already happening. Combat is reset whenever an NPC can't find their
target or they have none for whatever reason.