Add locks to doors
Doors may have locks. A lock is keyed by assigning the id of an item.
If that item is in the player’s inventory (or wielded), then the lock
can be unlocked. This allows players with the key to automatically
unlock and open doors as they move, preserving the existing behavior.

If a door has a lock and no key is assigned, then the lock is
considered to be unable to be unlocked.